requirements-dev.txt
-r requirements.txt

# Type checking
mypy==1.15.0
django-stubs[compatible-mypy]==5.2.0

# Linting
flakeheaven==3.3.0
flake8-docstrings==1.7.0

# Testing
pytest==8.3.5
pytest-django==4.9.0
factory-boy==3.3.3
